Extract DesktopIcon component
